scala 2.13.2

+++ sjvm/src/test/scala/com/github/thebridsk/utilities/logging/test/SystemTimeJvm.scala
@@ -12,7 +12,7 @@ object SystemTimeJvm {
 
     val st = new SystemTime {
       def currentTimeMillis() = {
-        System.currentTimeMillis()
+        System.currentTimeMillis().toDouble
       }
 
       val fmt = new SimpleDateFormat("HH:mm:ss.SSS")
